+++ GameEngine/src/com/javagameengine/scene/component/MeshRenderer.java
@@ -73,7 +73,7 @@ public class MeshRenderer extends RenderableComponent
 			material.create();
 	}
 
-	public void onUpdate(int delta)
+	public void onUpdate(float delta)
 	{
 		// Update mesh bounds and vertex buffers... if needed!
 	}
